+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_es.dart
@@ -387,4 +387,12 @@ No es posible añadir más de $limit archivos adjuntos
 
   @override
   String get linkDisabledError => 'Los enlaces están deshabilitados';
+
+  @override
+  String unreadMessagesSeparatorText(int unreadCount) {
+    if (unreadCount == 1) {
+      return '1 mensaje no leído';
+    }
+    return '$unreadCount mensajes no leídos';
+  }
 }